cs-drf-customers/requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
coreapi==2.3.3
coreschema==0.0.4
Django==3.2.9
django-filter==22.1
djangorestframework==3.12.4
Faker==0.7.4
psycopg2==2.9.2
psycopg2-binary==2.8.6
ptvsd==4.3.2
requests==2.26.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
validate-docbr==1.10.0
